requirements.txt
asgiref==3.8.1
black==25.1.0
boto3==1.37.31
botocore==1.37.31
click==8.1.8
Django==5.2
django-cuser==2017.3.16
djangorestframework==3.16.0
hashids==1.3.1
jmespath==1.0.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
python-dateutil==2.9.0.post0
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.3.0
user-agents==2.2.0
